Don suggestion 3 - move fixtures closer to use
